% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorFuns.R
\name{cleanColorGroup}
\alias{cleanColorGroup}
\title{cleanColorGroup}
\usage{
cleanColorGroup(colors, colorSet, splitchar = c(" ", "_"))
}
\description{
Function to split composite colors into elements consisting only of
elements in the second argument, colorSet.
}
