% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modomaDists.R
\name{cleanRHS}
\alias{cleanRHS}
\title{cleanRHS}
\usage{
cleanRHS(rhsvec, rhs)
}
\description{
Function cleanRHS removes all non-RHS items from the input vector
RHS codes are basically a combination of an optional letter (or
even two), a number of 1-3 digits, and a letter. Sometimes two
codes are indicated as 56 C-D
}
