% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorFuns.R
\name{countCorrect}
\alias{countCorrect}
\title{countCorrect}
\usage{
countCorrect(xtable)
}
\description{
The following function counts the number of cases where the row
label matches the column label. Row labels may contain two
elements, both are counted as correct.
}
