% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modomaDists.R
\name{getClosestTraits}
\alias{getClosestTraits}
\title{getClosest}
\usage{
getClosestTraits(indat, traitValues = NULL, n = 1)
}
\description{
Get n flowers closest to reference flower based on
selected subset of traits.
}
