% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorFuns.R
\name{getRMSquantile}
\alias{getRMSquantile}
\title{getRMSquantile}
\usage{
getRMSquantile(
  RHS,
  imagename,
  colTable,
  cspace = c("rgb", "HCL"),
  quant = 0.25,
  minInt = 5,
  plot = FALSE
)
}
\description{
compare individual pixel color values with the ground truth, and
return the specified quantile
}
