% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorFuns.R
\name{mainCol}
\alias{mainCol}
\title{mainCol}
\usage{
mainCol(
  densObj,
  what = c("count", "both", "bright"),
  hierClust = TRUE,
  hierCutree = 1.5,
  include,
  exclude,
  plot = FALSE,
  colorSpace = c("RGB", "HCL"),
  ...
)
}
\description{
Given a colour histogram, return the main colour. We can restrict
the choice: include takes precedence over exclude. Returns the
number of the main color.
}
