% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modomaDists.R
\name{modocol2rgb}
\alias{modocol2rgb}
\title{modocol2rgb}
\usage{
modocol2rgb(colstr, sepchars = c(" ", "_"))
}
\description{
standard R function col2rgb gives the translation for named colours
in R - several color indications in the naktuinbouw data are
composite. If all elements are individual colours then the average
RGB is returned, if words like light or dark are present then the
lighter or darker shades of the corresponding colours are used.
}
\details{
modocol2rgb takes one string, and a vector of sepchars
}
\examples{
modocol2rgb(c("dark yellow dark purple red"))
modocol2rgb(c("dark yellow", "dark purple", "red", "dark whatever"))
modocol2rgb("dark yellow")

}
