% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modoma.R
\docType{package}
\name{modoma-package}
\alias{modoma}
\alias{modoma-package}
\title{modoma: Analysis of flower images}
\description{
The goal is to automatically obtain variety characteristics for roses and gerberas on the basis of (standardized) flower images. The package also contains functionality to compare varieties on the basis of these characteristics with database records.
}
\keyword{internal}
