% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modomaDists.R
\name{mydist.ordered}
\alias{mydist.ordered}
\title{mydist.ordered}
\usage{
\method{mydist}{ordered}(x, ...)
}
\description{
for ordered factors special care is needed in case NA values are
present. Some factor levels may be missing, but I do take them into
account here. This is one way to provide information about level
spacings: simply add a number of levels in between that are not
used. Missing values are replaced by the maximum possible
distance. In an ordered factor with three levels, for example, the
maximum distance of the middle level to an NA is 1, the maximum
distance of the two extreme levels to an NA is 2. One could make a
case of imputing by half the maximum distance, which can be done by
simply dividing all elements in imputevec by 2. Let's see...
}
