% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorFuns.R
\name{pal}
\alias{pal}
\title{pal}
\usage{
pal(
  col,
  heights,
  border = "light gray",
  colSpace = c("rgb", "RGB", "HLS", "HSV", "LAB", "LUV", "polarLAB", "polarLUV"),
  mainCol = NULL,
  labels = is.null(mainCol),
  ...
)
}
\description{
Function to show a colour histogram, obtained by, e.g., clustering
image pixels. The first argument is an argument from class color,
(from colorspace) with color definition in a three-column matrix in
the coords element. Argument colSpace should be one
of the color-generating functions from package colorspace, or the
rgb() function. If the former, the color matrix will be transformed
into rgb colors. A bit clumsy, but the result of using two
different color representations: the S4 one from colorspace, and
the one from the regular graphics system.
}
