% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorFuns.R
\name{rgb2idx}
\alias{rgb2idx}
\title{rgb2idx}
\usage{
rgb2idx(rgb, colTable)
}
\description{
Function rgb2idx takes a matrix of RGB codes, calculates the Euclidean
distance to all RGB triplets in the color table, and returns
the closest one
}
