% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorFuns.R
\name{segmentImage}
\alias{segmentImage}
\title{segmentImage}
\usage{
segmentImage(
  imagemat,
  cspace = "RGB",
  minInt = 5,
  nsamp = 1000,
  quiet = FALSE,
  segmentImage = TRUE,
  fast = TRUE,
  ...
)
}
\description{
Function to return the main segmented colors in an image.
Background pixels are supposed to be black, with an
intensity not more than minInt. To fit the mixture model we use
sampling. If the model is to be fitted on the complete data then
nsamp should be set to NULL.
}
\details{
segmentImage3: the function that does all, analysis in RGB or HCL
space, kmeans (fast) or mclust. The function always takes a 3D
array that will provide the dimensions variable. Optimised somewhat
for speed: we avoid doing things on the matrix as a whole if it is
needed only on the sample.
}
