% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorFuns.R
\name{showColTab}
\alias{showColTab}
\title{showColTab}
\usage{
showColTab(colTab, nr, nc, nrmax = 10, ncmax = 10, labels, cex = 1, ...)
}
\description{
Function to create a figure with coloured rectangles, one for every
colour in the colour table that is the first argument
}
