% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorFuns.R
\name{showFlowerResult}
\alias{showFlowerResult}
\title{showFlowerResult}
\usage{
showFlowerResult(
  floName,
  rgbSolution,
  hclSolution,
  dbcol,
  mainColRGB = NULL,
  mainColHCL = NULL,
  format = c("tif", "jpg")
)
}
\description{
Function showing the original flower image on the left, the color
histograms of RGB and HCL approaches in the middle, and the ground
truth on the right
}
